requirement.txt.txt
fastapi
uvicorn
pymongo
bcrypt
python-jose
passlib
